# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Lunme API",
    "version": "1.0.0",
    "description": "Swagger gerado automaticamente pelas rotas do Next.js"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
...
    "/api/internal/uploads/capa": {
      "post": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Realiza upload de arquivos",
        "description": "Envia um arquivo em formato multipart/form-data e retorna a URL final após o upload.",
        "tags": [
          "Interno - files and uploads"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "path",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Caminho opcional onde o arquivo será salvo",
            "example": "pacientes/123"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "Arquivo a ser enviado"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Upload realizado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Upload realizado com sucesso!"
                    },
                    "url": {
                      "type": "string",
                      "example": "https://meuservidor.com/uploads/pasta/arquivo.png"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Nenhum arquivo enviado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Nenhum arquivo enviado"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro interno no servidor",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Erro interno no servidor"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Retorna a URL pública da foto da capa do livro",
        "description": "Busca a ulr da capa do livro do usuário pelo ID fornecido.",
        "tags": [
          "Interno - files and uploads"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID do usuário",
            "example": "clvxyz123abc789"
          }
        ],
        "responses": {
          "200": {
            "description": "URL da imagem retornada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "url": {
                      "type": "string",
                      "example": "https://supabase.storage.com/bucket/capa/user123.png"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "ID do usuário não fornecido",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "ID do usuário não fornecido"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "capa não encontrado ou sem imagem cadastrada",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Usuário ou imagem não encontrada"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro interno no servidor",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Erro interno no servidor"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": []
}